+++ src/app/entities/service-group.entity.ts
@@ -1,0 +1,9 @@
+import { ServiceEntity } from "./service.entity";
+
+export interface ServiceGroupEntity {
+    id: string,
+    serviceIds: number[],
+    selected: boolean,
+    names: string[],
+    services?: ServiceEntity[],
+}